requirements.txt
dj-database-url==0.4.1
Django==1.8.13
djangorestframework==3.3.3
gunicorn==19.6.0
nose==1.3.7
pkg-resources==0.0.0
pycrypto==2.6.1
whitenoise==3.2
